feat: .handle_quick_operation of websocket
