+++ src/onebot11/action/go-cqhttp/QuickOperation.ts
@@ -1,0 +1,17 @@
+import BaseAction from '../BaseAction'
+import { handleQuickOperation, QuickOperation, QuickOperationEvent } from '../../server/quick-operation'
+import { log } from '@/common/utils'
+import { ActionName } from '../types'
+
+interface Payload{
+  context: QuickOperationEvent,
+  operation: QuickOperation
+}
+
+export class GoCQHTTHandleQuickOperation extends BaseAction<Payload, null>{
+  actionName = ActionName.GoCQHTTP_HandleQuickOperation
+  protected async _handle(payload: Payload): Promise<null> {
+    handleQuickOperation(payload.context, payload.operation).then().catch(log);
+    return null
+  }
+}
