+++ src/onebot11/server/ws/ReverseWebsocket.ts
@@ -4,7 +4,7 @@ import { ActionName } from '../../action/types'
 import { OB11Response } from '../../action/OB11Response'
 import BaseAction from '../../action/BaseAction'
 import { actionMap } from '../../action'
-import { postWsEvent, registerWsEventSender, unregisterWsEventSender } from '../postOB11Event'
+import { postWsEvent, registerWsEventSender, unregisterWsEventSender } from '../post-ob11-event'
 import { wsReply } from './reply'
 import { WebSocket as WebSocketClass } from 'ws'
 import { OB11HeartbeatEvent } from '../../event/meta/OB11HeartbeatEvent'
